Test: consolidate helpers and add barrel
