+++ test/helpers/index.ts
@@ -1,0 +1,4 @@
+export * from './diagnostics.js';
+export * from './cli.js';
+export * from './cliBuild.js';
+export * from './lowered_program.js';
